# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Health Hub API
  version: 0.20.0
  description: |
    API для приложения "Центр здоровья".
    Canonical file — все эндпоинты описаны здесь.

    v0.20.0: Added Food Preferences API (GET/POST/DELETE /v1/food/prefs) and Meal Plans API (GET/PUT/DELETE /v1/meal/plan, GET /v1/meal/today). Extended FeedDayResponse with meal_today, meal_plan_title, food_prefs_count.
    v0.19.0: Added Nutrition Targets API endpoints GET/PUT /v1/nutrition/targets and extended FeedDayResponse with nutrition_targets and nutrition_progress.
    v0.18.0: Added Workout Plans API endpoints and nutrition_plan proposal support.
    v0.17.0: Added supplement schedules API and vitamins_schedule proposal apply flow.
    v0.16.0: Added AI proposal management endpoints GET /v1/ai/proposals, POST /v1/ai/proposals/{id}/apply, POST /v1/ai/proposals/{id}/reject.
    v0.15.0: Added Chat MVP endpoint GET/POST /v1/chat/messages with assistant proposals.
    v0.14.0: Added user settings endpoints GET/PUT /v1/settings and notifications settings persistence.
    v0.13.0: Added Email OTP auth endpoints POST /v1/auth/email/request and POST /v1/auth/email/verify.
    v0.12.0: Added POST /v1/auth/siwa with Apple identity token verification via JWKS.
    v0.11.0: Added POST /v1/auth/dev for local dev JWT issuance.
    v0.10.0: Added Sign in with Apple (SIWA) + JWT authentication, Intakes (Water & Supplements) API.
    v0.9.0: Added Inbox/Notifications API for server-side notifications.
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /v1/inbox:
    get:
      summary: List notifications
      description: Получение списка уведомлений для профиля
      operationId: listNotifications
      parameters:
        - name: profile_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: only_unread
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Список уведомлений
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/food/prefs:
    get:
      summary: List food preferences
      description: Получить список обычных продуктов с возможностью поиска
      operationId: listFoodPrefs
      tags:
        - Food Preferences
      parameters:
        - name: profile_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          required: false
          description: Поисковый запрос (по названию)
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Список продуктов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFoodPrefsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Неавторизован
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create or update food preference
      description: Создать или обновить продукт
      operationId: upsertFoodPref
      tags:
        - Food Preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertFoodPrefRequest"
      responses:
        "200":
          description: Продукт создан или обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodPrefDTO"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Неавторизован
        "409":
          description: Продукт с таким названием уже существует или достигнут лимит (200)
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        profile_id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
            [
              low_sleep,
              low_activity,
              missing_morning_checkin,
              missing_evening_checkin,
            ]
        title:
          type: string
        body:
          type: string
        source_date:
          type: string
          format: date
          description: "Дата, к которой относится уведомление"
        severity:
          type: string
          enum: [info, warn]
        created_at:
          type: string
          format: date-time
        read_at:
          type: string
          format: date-time
      required: [id, profile_id, kind, title, body, severity, created_at]
    InboxListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
      required: [notifications]
...
    FoodPrefDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        profile_id:
          type: string
          format: uuid
        name:
          type: string
          description: Название продукта
        tags:
          type: array
          items:
            type: string
          description: Теги (например, "fruit", "protein", "breakfast")
        kcal_per_100g:
          type: integer
          minimum: 0
          description: Калории на 100г
        protein_g_per_100g:
          type: integer
          minimum: 0
          description: Белки на 100г (г)
        fat_g_per_100g:
          type: integer
          minimum: 0
          description: Жиры на 100г (г)
        carbs_g_per_100g:
          type: integer
          minimum: 0
          description: Углеводы на 100г (г)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - profile_id
        - name
        - tags
        - kcal_per_100g
        - protein_g_per_100g
        - fat_g_per_100g
        - carbs_g_per_100g
        - created_at
        - updated_at
    ListFoodPrefsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FoodPrefDTO"
        total:
          type: integer
          description: Общее количество продуктов (без учёта offset)
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    UpsertFoodPrefRequest:
      type: object
      properties:
        profile_id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Название продукта
        tags:
          type: array
          items:
            type: string
          description: Теги
        kcal_per_100g:
          type: integer
          minimum: 0
          default: 0
        protein_g_per_100g:
          type: integer
          minimum: 0
          default: 0
        fat_g_per_100g:
          type: integer
          minimum: 0
          default: 0
        carbs_g_per_100g:
          type: integer
          minimum: 0
          default: 0
      required:
        - profile_id
        - name
        - tags

    # --- Meal Plans ---
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Код ошибки
              example: "invalid_range"
            message:
              type: string
              description: Описание ошибки
              example: "Date range exceeds maximum of 90 days"
          required: [code, message]
      required: [error]
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен, полученный через POST /v1/auth/dev, /v1/auth/siwa или /v1/auth/email/verify.
        При AUTH_MODE=none токен не обязателен; при наличии токена сервер валидирует его.
        Header: Authorization: Bearer <jwt>

# По умолчанию аутентификация НЕ применяется (AUTH_MODE=none).
# Когда включится строгая защита, добавить security на защищённые эндпоинты:
# security:
#   - BearerAuth: []